requirements.txt
web3==6.11.3
rich==13.7.0
textual==0.44.1
click==8.1.7
python-dotenv==1.0.0
requests==2.31.0
setuptools>=65.0.0
pytest==7.4.3
pytest-mock==3.12.0
pre-commit==3.5.0
black==23.11.0
flake8==6.1.0